Reduced code duplication by creating an abstract class that can be used for creating a "SubmitterExecutorInterface" implementation. This reduces code now, but will be more useful in the future.
